+++ src/decorators/index.ts
@@ -4,3 +4,4 @@
 export { BotDeclaration } from "./BotDeclaration";
 export { ConnectorDeclaration } from "./ConnectorDeclaration";
 export { OutgoingWebhookDeclaration } from "./OutgoingWebhookDeclaration";
+export { MessageExtensionDeclaration } from "./MessageExtensionDeclaration";