+++ src/impl/index.ts
@@ -1,0 +1,1 @@
+export { TeamsBot } from "./TeamsBot";